// src/lib2/models/DetailEcritureDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DetailEcritureDto = {
    id?: string;
    libelle: string;
    sens: string;
    lettree?: boolean;
    pointee?: boolean;
    notes?: string;
    ecritureComptableId: string;
    compteComptableId: string;
    montantDebit?: number;
    montantCredit?: number;
    dateLettrage?: string;
    referenceBancaire?: string;
    dateEcriture?: string;
};